update : change the plugin generator files to separate folder
