--- packages/generator/ni_measurement_plugin_sdk_generator/__main__.py
@@ -1,5 +1,0 @@
-""""Creates a measurement through use of template.py."""
-
-from ni_measurement_plugin_sdk_generator.template import create_measurement
-
-create_measurement()
