+++ packages/generator/ni_measurement_plugin_sdk_generator/ni_measurement_plugin_generator/__main__.py
@@ -1,0 +1,5 @@
+""""Creates a measurement through use of template.py."""
+
+from ni_measurement_plugin_sdk_generator.ni_measurement_plugin_generator.template import create_measurement
+
+create_measurement()
