+++ packages/generator/ni_measurement_plugin_sdk_generator/ni_measurement_plugin_generator/templates/start.bat.mako
